runner/requirements.txt
anthropic>=0.40.0
openai>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
click>=8.0.0
python-dotenv>=1.0.0
tomli>=2.0.0;python_version<"3.11"
qdrant-client>=1.7.0
duckduckgo-search>=6.0.0
